feat: migrate ts rules

+++ scripts/migrate-ts-eslint.ts
@@ -18,9 +18,28 @@ const root = fileURLToPath(new URL('../../typescript-eslint', import.meta.url))
 const targetRoot = fileURLToPath(new URL('../packages/eslint-plugin-ts', import.meta.url))
 const target = join(targetRoot, 'rules')
 
-// TOD: add more rules, and wait for the official deprecation list
+// Wait for the official deprecation list
+// Currently listing the rules that marked as 'type: layout' in the TypeScript ESLint repo
 const names = [
+  'block-spacing',
+  'brace-style',
+  'comma-dangle',
+  'comma-spacing',
+  'func-call-spacing',
+  'indent',
+  'key-spacing',
+  'keyword-spacing',
+  'lines-around-comment',
+  'lines-between-class-members',
+  'member-delimiter-style',
+  'no-extra-parens',
+  'padding-line-between-statements',
+  'quotes',
   'semi',
+  'space-before-blocks',
+  'space-before-function-paren',
+  'space-infix-ops',
+  'type-annotation-spacing',
 ]
 
 async function run() {
@@ -50,23 +69,34 @@ async function run() {
           '\'../util',
           '\'../../util',
         )
+        .replaceAll(
+          'eslint-disable-next-line deprecation/deprecation',
+          '',
+        )
       // js = `// @ts-check\n${js}`
       await fs.writeFile(join(target, name, `${name}.ts`), js, 'utf-8')
 
       const md = await fs.readFile(join(root, 'packages/eslint-plugin/docs/rules', `${name}.md`), 'utf-8')
       await fs.writeFile(join(target, name, 'README.md'), md, 'utf-8')
 
-      let test = await fs.readFile(join(root, 'packages/eslint-plugin/tests/rules', `${name}.test.ts`), 'utf-8')
-      test = test
-        .replaceAll(
-          '\'../../src/rules/',
-          '\'./',
-        )
-        .replaceAll(
-          'eslint "@typescript-eslint/internal',
-          '',
-        )
-      await fs.writeFile(join(target, name, `${name}.test.ts`), test, 'utf-8')
+      const testPath = join(root, 'packages/eslint-plugin/tests/rules', `${name}.test.ts`)
+      if (existsSync(testPath)) {
+        let test = await fs.readFile(testPath, 'utf-8')
+        test = test
+          .replaceAll(
+            '\'../../src/rules/',
+            '\'./',
+          )
+          .replaceAll(
+            'eslint "@typescript-eslint/internal',
+            '',
+          )
+          .replaceAll(
+            './indent/utils',
+            '../indent/utils',
+          )
+        await fs.writeFile(join(target, name, `${name}.test.ts`), test, 'utf-8')
+      }
     }),
   )
 }
